Rename FeathersAppInfo to FeathersApp
