+++ src/index.js
@@ -4,7 +4,7 @@ import React from "react";
 import FeathersAppProvider from "./components/FeathersAppProvider";
 import FeathersQuery from "./components/FeathersQuery";
 import FeathersGet from "./components/FeathersGet";
-import FeathersAppInfo from "./components/FeathersAppInfo";
+import FeathersApp from "./components/FeathersApp";
 
 const ASC = 1;
 const DESC = -1;
@@ -13,7 +13,7 @@ export {
   FeathersAppProvider,
   FeathersQuery,
   FeathersGet,
-  FeathersAppInfo,
+  FeathersApp,
   ASC,
   DESC
 };
